fix: move dataset to src
