+++ apps/prisma-postgres/src/controllers/province/updateProvince.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { IProvinceDTO } from '../../../../dataset/location.dto';
+import { IProvinceDTO } from '../../../dataset/location.dto';
 import updateProvinceService from '../../services/province/updateProvince.service';
 
 const updateProvince = async (req: Request, res: Response) => {
